Removing autocast.

+++ src/diffusers/models/unet_2d_condition.py
@@ -224,6 +224,7 @@ class UNet2DConditionModel(ModelMixin, ConfigMixin):
         t_emb = self.time_proj(timesteps)
         emb = self.time_embedding(t_emb)
 
+        sample = sample.to(torch.float16)
         # 2. pre-process
         sample = self.conv_in(sample)
 
@@ -260,7 +261,8 @@ class UNet2DConditionModel(ModelMixin, ConfigMixin):
         # 6. post-process
         # make sure hidden states is in float32
         # when running in half-precision
-        sample = self.conv_norm_out(sample.float()).type(sample.dtype)
+        # sample = self.conv_norm_out(sample.float()).type(sample.dtype)
+        sample = self.conv_norm_out(sample).type(sample.dtype)
         sample = self.conv_act(sample)
         sample = self.conv_out(sample)
